Initial commit for R_BbiReader
An R interface to BigWig and BigBed files.

The C++ code in src/ can be compiled using R CMD commands using
the Makevars file to create a shared library that can be used by
R to access data in BigWig files using the functions in the R/
directory. Currently BigBed files are not supported but it should
be straightforward to add support for bigBed as well.

The C++ code can also be compiled using the Makefile and make (though
the Makefile will probably need to be modified) to create binary that
can be used to test the different classes. The main() for this is in
browse_header_information.cpp. This can be used to check performance
on different systems (edit the main() function to suit your needs).