Added partial support for BigBed files
Added support for BigBed files through the BigFile object. This is still
a little experimental, and needs to be checked more carefully.

The BigBed support is still incomplete, as the format can support an
arbitrary fields in arbitrary orders.

Also modified BigFile to read the summary if a summary offset is defined.